Decode params before injecting them into `query`
Fixes #145

When urls are constructed in the `Link` component, all params are URI
encoded so that they won't break routing. When params are read however,
they are never decoded. This leads to an inconsistent behavior between
the server and the client when params are passed down the page component
as `query`. On the server, they come down untouched (encoded). On the
client, they are not encoded.

Consider this link:

  <Link route="foo" params={{ b: 'f/g' }}><a>Foo</a></Link>

and this page component

  FooPage.getInitialProps = async ({ query }) => {
    console.log(query.b);
  }

On the server, this page would log "f%2Fg" when navigated to. On the
client it would log "f/g".

By decoding params before they are injected into the `query` params, we
get the same behavior on the client and the server. We also make the
behavior consistent with how regular query params work.